Add JSON converter and gameSessionCardId to Card class  Added JSON converter and gameSessionCardId for card identification Implemented ListToJsonConverter to handle JSON conversion for lists. Implemented ListToJsonConverter to handle JSON conversion for lists. Updated Card class to include gameSessionCardId for session-specific card identification. Modified GameSession to utilize gameSessionCardId during card distribution. Updated Card class to include gameSessionCardId for identifying cards in each session. Enhanced GameSession to initialize and distribute cards with gameSessionCardId. Updated GameController with endpoints for testing and distributing cards.n
