+++ mindbug-backend/src/main/java/com/mindbug/models/Card.java
@@ -13,7 +13,6 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 @Entity
 public class Card {
-    //private int number;
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -27,7 +26,7 @@ public class Card {
     private List<String> keywords;
     private String trigger_condition;
     @ElementCollection
-    private List<EffectType> effect_type;
+    private List<EffectType> effect;
 
     public Card() {
     }
@@ -39,7 +38,7 @@ public class Card {
         this.power = card.power;
         this.keywords = card.keywords;
         this.trigger_condition = card.trigger_condition;
-        this.effect_type = card.effect_type;
+        this.effect = card.effect;
     }
 
     public GameSessionCard getGameSessionCardId() {
@@ -90,12 +89,11 @@ public class Card {
         this.trigger_condition = triggerCondition;
     }
 
-    public List<EffectType> getEffect_type() {
-        return effect_type;
+    public List<EffectType> getEffect() {
+        return effect;
     }
-
-    public void setEffect_type(List<EffectType> effectTypes) {
-        this.effect_type = effectTypes;
+    public void setEffect(List<EffectType> effect) {
+        this.effect = effect;
     }
     @Override
     public String toString() {
@@ -106,106 +104,105 @@ public class Card {
                 ", power=" + power +
                 ", keywords=" + keywords +
                 ", trigger_condition='" + trigger_condition + '\'' +
-                ", effect_type=" + effect_type +
+                ", effect=" + effect +
                 '}';
     }
 }
-
-@Embeddable
-class EffectType {
-    private String type;
-    private String target;
-
-    @JsonProperty("effectValue")
-    private int effectValue;
-
-    private String action;
-    @Convert(converter = ConditionConverter.class)
-    private Map<String, Object> condition;
-    private int amount;
-    private String operation;
-    @Convert(converter = ListToJsonConverter.class)
-    private List<String> keywords;
-
-    // Getters and setters
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    public String getTarget() {
-        return target;
-    }
-
-    public void setTarget(String target) {
-        this.target = target;
-    }
-
-    public int getEffectValue() {
-        return effectValue;
-    }
-
-    public void setEffectValue(int effectValue) {
-        this.effectValue = effectValue;
-    }
-
-    public String getAction() {
-        return action;
-    }
-
-    public void setAction(String action) {
-        this.action = action;
-    }
-
-    public Map<String, Object> getCondition() {
-        return condition;
-    }
-
-    public void setCondition(Map<String, Object> condition) {
-        this.condition = condition;
-    }
-
-    public int getAmount() {
-        return amount;
-    }
-
-    public void setAmount(int amount) {
-        this.amount = amount;
-    }
-
-    public String getOperation() {
-        return operation;
-    }
-
-    public void setOperation(String operation) {
-        this.operation = operation;
-    }
-
-    public List<String> getKeywords() {
-        return keywords;
-    }
-
-    public void setKeywords(List<String> keywords) {
-        this.keywords = keywords;
-    }
-
-    @Override
-    public String toString() {
-        return "EffectType{" +
-                "type='" + type + '\'' +
-                ", target='" + target + '\'' +
-                ", effectValue=" + effectValue +
-                ", action='" + action + '\'' +
-                ", condition=" + condition +
-                ", amount=" + amount +
-                ", operation='" + operation + '\'' +
-                ", keywords=" + keywords +
-                '}';
-    }
-}
+//
+//@Embeddable
+//class Effect_Type {
+//    private String type;
+//    private String target;
+//
+//    @JsonProperty("effectValue")
+//    private int effectValue;
+//
+//    private String action;
+//    @Convert(converter = ConditionConverter.class)
+//    private Map<String, Object> condition;
+//    private int amount;
+//    private String operation;
+//    @Convert(converter = ListToJsonConverter.class)
+//    private List<String> keywords;
+//
+//    // Getters and setters
+//
+//    public String getType() {
+//        return type;
+//    }
+//
+//    public void setType(String type) {
+//        this.type = type;
+//    }
+//
+//    public String getTarget() {
+//        return target;
+//    }
+//
+//    public void setTarget(String target) {
+//        this.target = target;
+//    }
+//
+//    public int getEffectValue() {
+//        return effectValue;
+//    }
+//
+//    public void setEffectValue(int effectValue) {
+//        this.effectValue = effectValue;
+//    }
+//
+//    public String getAction() {
+//        return action;
+//    }
+//
+//    public void setAction(String action) {
+//        this.action = action;
+//    }
+//
+//    public Map<String, Object> getCondition() {
+//        return condition;
+//    }
+//
+//    public void setCondition(Map<String, Object> condition) {
+//        this.condition = condition;
+//    }
+//
+//    public int getAmount() {
+//        return amount;
+//    }
+//
+//    public void setAmount(int amount) {
+//        this.amount = amount;
+//    }
+//
+//    public String getOperation() {
+//        return operation;
+//    }
+//
+//    public void setOperation(String operation) {
+//        this.operation = operation;
+//    }
+//
+//    public List<String> getKeywords() {
+//        return keywords;
+//    }
+//
+//    public void setKeywords(List<String> keywords) {
+//        this.keywords = keywords;
+//    }
+//
+//    @Override
+//    public String toString() {
+//        return "EffectType{" +
+//                "type='" + type + '\'' +
+//                ", target='" + target + '\'' +
+//                ", effectValue=" + effectValue +
+//                ", action='" + action + '\'' +
+//                ", condition=" + condition +
+//                ", amount=" + amount +
+//                ", operation='" + operation + '\'' +
+//                ", keywords=" + keywords +
+//                '}';
+//    }
