+++ mindbug-backend/src/main/java/com/mindbug/models/Keyword.java
@@ -1,0 +1,9 @@
+package com.mindbug.models;
+
+public enum Keyword {
+    POISONOUS,
+    SNEAKY,
+    HUNTER,
+    TOUGH,
+    FRENZY
+}
